go fmt, and some formatting tweaks
when using fatal it wasn't appending a newline, so messages were confusing
this fixes that.
also, add millisecond-level timing for how long it takes us to respond
to requests. The *data* is available in UnixNano() but it requires
doing the subtraction somehow, and also misses some of the overhead
as the 'received challenge' only triggers after we've processed
some of the request.